requirements.txt
aiogram==3.4.1
aiosqlite
requests
python-dotenv